Gael UltraFeat: Finished all frontend and backend functionalities except for addEditRental and front-end for confirming orders

+++ software/src/main/java/backend/software/dto/editAppointment.java
@@ -1,22 +1,24 @@
 package backend.software.dto;
 
+import java.math.BigDecimal;
+
 public class editAppointment {
     private String customerName;
     private String category;
     private String description;
     private String dateTimeAppointed;
-    private Double cost;
+    private BigDecimal cost;
     private String name;
     private Long id;
     private Boolean ongoing;
     
     public editAppointment(){}
 
-    public Double getCost() {
+    public BigDecimal getCost() {
         return cost;
     }
 
-    public void setCost(Double cost) {
+    public void setCost(BigDecimal cost) {
         this.cost = cost;
     }
